Move active plants endpoint to echo controller
