+++ backend/api/Controllers/EchoController.cs
@@ -1,7 +1,9 @@
-﻿using System.Text.Json;
+﻿using System.Globalization;
+using System.Text.Json;
 using Api.Controllers.Models;
 using Api.Services;
 using Api.Services.Models;
+using Api.Utilities;
 using Microsoft.AspNetCore.Authorization;
 using Microsoft.AspNetCore.Mvc;
 
@@ -16,10 +18,13 @@ public class EchoController : ControllerBase
 
     private readonly IEchoService _echoService;
 
-    public EchoController(ILogger<EchoController> logger, IEchoService echoService)
+    private readonly IRobotService _robotService;
+
+    public EchoController(ILogger<EchoController> logger, IEchoService echoService, IRobotService robotService)
     {
         _logger = logger;
         _echoService = echoService;
+        _robotService = robotService;
     }
 
     /// <summary>
@@ -152,6 +157,51 @@ public class EchoController : ControllerBase
         try
         {
             var echoPlantInfos = await _echoService.GetEchoPlantInfos();
+            return Ok(echoPlantInfos);
+        }
+        catch (HttpRequestException e)
+        {
+            _logger.LogError(e, "Error getting plant info from Echo");
+            return new StatusCodeResult(StatusCodes.Status502BadGateway);
+        }
+        catch (JsonException e)
+        {
+            _logger.LogError(e, "Error deserializing plant info response from Echo");
+            return new StatusCodeResult(StatusCodes.Status500InternalServerError);
+        }
+    }
+
+    /// <summary>
+    /// Get all plants associated with an active robot.
+    /// </summary>
+    /// <remarks>
+    /// <para> Retrieves the plants that have an active robot </para>
+    /// </remarks>
+    [HttpGet]
+    [Authorize(Roles = Role.User)]
+    [Route("active-plants")]
+    [ProducesResponseType(typeof(IList<EchoPlantInfo>), StatusCodes.Status200OK)]
+    [ProducesResponseType(StatusCodes.Status400BadRequest)]
+    [ProducesResponseType(StatusCodes.Status401Unauthorized)]
+    [ProducesResponseType(StatusCodes.Status403Forbidden)]
+    [ProducesResponseType(StatusCodes.Status500InternalServerError)]
+    public async Task<ActionResult<IList<EchoPlantInfo>>> GetActivePlants()
+    {
+        var plants = await _robotService.ReadAllActivePlants();
+
+        if (plants == null)
+        {
+            _logger.LogWarning("Could not retrieve robot plants information");
+            throw new RobotInformationNotAvailableException("Could not retrieve robot plants information");
+        }
+
+        plants = plants.Select(p => p.ToLower(CultureInfo.CurrentCulture));
+
+        try
+        {
+            var echoPlantInfos = await _echoService.GetEchoPlantInfos();
+
+            echoPlantInfos = echoPlantInfos.Where(p => plants.Contains(p.PlantCode.ToLower(CultureInfo.CurrentCulture))).ToList();
             return Ok(echoPlantInfos);
         }
         catch (HttpRequestException e)
