+++ backend/api/Controllers/RobotController.cs
@@ -281,51 +281,6 @@ public class RobotController : ControllerBase
     }
 
     /// <summary>
-    /// Get all plants associated with an active robot.
-    /// </summary>
-    /// <remarks>
-    /// <para> Retrieves the plants that have an active robot </para>
-    /// </remarks>
-    [HttpGet]
-    [Authorize(Roles = Role.User)]
-    [Route("active-plants")]
-    [ProducesResponseType(typeof(IList<EchoPlantInfo>), StatusCodes.Status200OK)]
-    [ProducesResponseType(StatusCodes.Status400BadRequest)]
-    [ProducesResponseType(StatusCodes.Status401Unauthorized)]
-    [ProducesResponseType(StatusCodes.Status403Forbidden)]
-    [ProducesResponseType(StatusCodes.Status500InternalServerError)]
-    public async Task<ActionResult<IList<EchoPlantInfo>>> GetActivePlants()
-    {
-        var plants = await _robotService.ReadAllActivePlants();
-
-        // TODO: call echo API here to get the remaining info so that we can rely solely on this endpoint
-        // So replace where we call getEchoPlantInfo with getActivePlants, based on a toggle
-        if (plants == null)
-        {
-            _logger.LogWarning("Could not retrieve robot plants information");
-            throw new RobotInformationNotAvailableException("Could not retrieve robot plants information");
-        }
-
-        try
-        {
-            var echoPlantInfos = await _echoService.GetEchoPlantInfos();
-
-            echoPlantInfos = echoPlantInfos.Where(p => plants.Contains(p.PlantCode)).ToList();
-            return Ok(echoPlantInfos);
-        }
-        catch (HttpRequestException e)
-        {
-            _logger.LogError(e, "Error getting plant info from Echo");
-            return new StatusCodeResult(StatusCodes.Status502BadGateway);
-        }
-        catch (JsonException e)
-        {
-            _logger.LogError(e, "Error deserializing plant info response from Echo");
-            return new StatusCodeResult(StatusCodes.Status500InternalServerError);
-        }
-    }
-
-    /// <summary>
     /// Add a video stream to a given robot
     /// </summary>
     /// <remarks>
